+++ b4n-tui/widgets/footer.rs
@@ -5,10 +5,15 @@ use ratatui::layout::{Constraint, Direction, Flex, Layout, Margin, Rect};
 use ratatui::style::Style;
 use ratatui::text::{Line, Span};
 use ratatui::widgets::{Block, Paragraph};
+use std::collections::VecDeque;
 use std::{rc::Rc, time::Instant};
 use tokio::sync::mpsc::{self, UnboundedReceiver};
 
+use crate::widgets::history::{BottomPane, MessageItem};
+use crate::{ResponseEvent, Responsive, TuiEvent};
+
 const FOOTER_APP_VERSION: &str = concat!(" b4n v", env!("CARGO_PKG_VERSION"), " ");
+const MESSAGE_HISTORY_SIZE: usize = 20;
 
 /// Footer widget.
 pub struct Footer {
@@ -18,9 +23,13 @@ pub struct Footer {
     message: Option<Notification>,
     messages_rx: UnboundedReceiver<Notification>,
     message_received_time: Instant,
+    message_history: VecDeque<(usize, Instant, Notification)>,
+    message_count: usize,
     icons: Vec<Icon>,
     icons_rx: UnboundedReceiver<IconAction>,
     notifications_tx: NotificationSink,
+    history_pane: Option<BottomPane>,
+    area: Rect,
 }
 
 impl Default for Footer {
@@ -37,9 +46,13 @@ impl Default for Footer {
             message: None,
             messages_rx,
             message_received_time: Instant::now(),
+            message_history: VecDeque::with_capacity(MESSAGE_HISTORY_SIZE),
+            message_count: 0,
             icons: Vec::new(),
             icons_rx,
             notifications_tx,
+            history_pane: None,
+            area: Rect::default(),
         }
     }
 }
@@ -69,11 +82,33 @@ impl Footer {
             .split(area)
     }
 
+    /// Returns `true` if footer is showing history pane at the moment.
+    pub fn is_message_history_visible(&self) -> bool {
+        self.history_pane.is_some()
+    }
+
+    /// Shows history pane.
+    pub fn show_message_history(&mut self) {
+        if self.history_pane.is_none() {
+            self.history_pane = Some(BottomPane::new(self.get_history_messages().into()))
+        }
+    }
+
+    /// Hides history pane.
+    pub fn hide_message_history(&mut self) {
+        if self.history_pane.is_some() {
+            self.history_pane = None;
+        }
+    }
+
     /// Draws [`Footer`] on the provided frame area.
     pub fn draw(&mut self, frame: &mut Frame<'_>, area: Rect, theme: &Theme) {
+        self.area = area;
         self.draw_footer(frame, area, theme);
 
-        if self.has_message_to_show()
+        self.update_current_message();
+        if !self.is_message_history_visible()
+            && self.has_message_to_show()
             && let Some(message) = &self.message
         {
             let [area] = Layout::horizontal([Constraint::Length(message.text.chars().count() as u16)])
@@ -81,6 +116,13 @@ impl Footer {
                 .areas(area.inner(Margin::new(2, 0)));
             frame.render_widget(Footer::get_message(&message.text, message.is_error, &theme.colors), area);
         }
+    }
+
+    /// Draws messages history on the bottom of the specified area.
+    pub fn draw_history(&mut self, frame: &mut Frame<'_>, area: Rect, theme: &Theme) {
+        if let Some(pane) = &mut self.history_pane {
+            pane.draw(frame, area, theme);
+        };
     }
 
     fn draw_footer(&mut self, frame: &mut Frame<'_>, area: Rect, theme: &Theme) {
@@ -118,7 +160,6 @@ impl Footer {
 
     /// Returns `true` if there is a message to show.
     fn has_message_to_show(&mut self) -> bool {
-        self.update_current_message();
         if let Some(message) = &self.message {
             if self.message_received_time.elapsed().as_millis() <= u128::from(message.duration) {
                 true
@@ -186,12 +227,28 @@ impl Footer {
     fn update_current_message(&mut self) {
         let mut message = None;
         while let Ok(current) = self.messages_rx.try_recv() {
+            if self.message_history.len() >= MESSAGE_HISTORY_SIZE {
+                self.message_history.pop_back();
+            }
+
+            self.message_history
+                .push_front((self.message_count, Instant::now(), current.clone()));
+            self.message_count = self.message_count.overflowing_add(1).0;
             message = Some(current);
         }
 
         if message.is_some() {
-            self.message = message;
-            self.message_received_time = Instant::now();
+            if message.as_ref().is_some_and(|m| m.duration > 0) {
+                self.message = message;
+                self.message_received_time = Instant::now();
+            }
+
+            if self.history_pane.is_some() {
+                let new_messages = self.get_history_messages();
+                if let Some(pane) = &mut self.history_pane {
+                    pane.update(new_messages);
+                }
+            }
         }
     }
 
@@ -242,5 +299,29 @@ impl Footer {
 
         spans.push(Span::styled(" ".repeat(width.saturating_sub(total)), &colors.footer.text));
         Line::from(spans)
+    }
+
+    fn get_history_messages(&self) -> Vec<MessageItem> {
+        self.message_history
+            .iter()
+            .map(|(c, t, n)| MessageItem::from(n, *t, *c))
+            .collect()
+    }
+}
+
+impl Responsive for Footer {
+    fn process_event(&mut self, event: &TuiEvent) -> ResponseEvent {
+        if let Some(pane) = &mut self.history_pane {
+            if pane.process_event(event) == ResponseEvent::Cancelled {
+                self.history_pane = None;
+            }
+
+            return ResponseEvent::Handled;
+        } else if event.is_in(crate::MouseEventKind::LeftClick, self.area) {
+            self.show_message_history();
+            return ResponseEvent::Handled;
+        }
+
+        ResponseEvent::NotHandled
     }
 }